progress proyek 2

+++ src/api/songs/handler.js
@@ -15,17 +15,16 @@ class SongsHandler {
   async postSongHandler(request, h) {
     try {
       this._validator.validateSongPayload(request.payload);
-      const {
-        title = 'untitled', year, performer, genre, duration,
-      } = request.payload;
+      const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
       const songId = await this._service.addSong({
-        title, year, performer, genre, duration,
+        title, body, tags, owner: credentialId,
       });
 
       const response = h.response({
         status: 'success',
-        message: 'Musik berhasil ditambahkan',
+        message: 'Catatan berhasil ditambahkan',
         data: {
           songId,
         },
@@ -42,8 +41,9 @@ class SongsHandler {
         return response;
       }
 
+      // Server ERROR!
       const response = h.response({
-        status: 'fail',
+        status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
@@ -52,8 +52,9 @@ class SongsHandler {
     }
   }
 
-  async getSongsHandler() {
-    const songs = await this._service.getSongs();
+  async getSongsHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const songs = await this._service.getSongs(credentialId);
     return {
       status: 'success',
       data: {
@@ -65,6 +66,9 @@ class SongsHandler {
   async getSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       const song = await this._service.getSongById(id);
       return {
         status: 'success',
@@ -82,6 +86,7 @@ class SongsHandler {
         return response;
       }
 
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.',
@@ -95,18 +100,14 @@ class SongsHandler {
   async putSongByIdHandler(request, h) {
     try {
       this._validator.validateSongPayload(request.payload);
-      const {
-        title, year, performer, genre, duration,
-      } = request.payload;
-      const { id } = request.params;
+      const { id } = request.params; const { id: credentialId } = request.auth.credentials;
 
-      await this._service.editSongById(id, {
-        title, year, performer, genre, duration,
-      });
+      await this._service.verifySongOwner(id, credentialId);
+      await this._service.editSongById(id, request.payload);
 
       return {
         status: 'success',
-        message: 'Musik berhasil diperbarui',
+        message: 'Catatan berhasil diperbarui',
       };
     } catch (error) {
       if (error instanceof ClientError) {
@@ -118,6 +119,7 @@ class SongsHandler {
         return response;
       }
 
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.',
@@ -131,11 +133,14 @@ class SongsHandler {
   async deleteSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       await this._service.deleteSongById(id);
 
       return {
         status: 'success',
-        message: 'Musik berhasil dihapus',
+        message: 'Catatan berhasil dihapus',
       };
     } catch (error) {
       if (error instanceof ClientError) {
@@ -147,6 +152,7 @@ class SongsHandler {
         return response;
       }
 
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.',
